[Add] :  render function in Handler.php,  [UP]  :  permission improvement

+++ routes/api.php
@@ -30,15 +30,18 @@ Route::post('/login', [AuthController::class, 'login'])->name('login');
     // todo add middleware to check if user is registered, removed for developement reasons
 Route::prefix('/dashboard')->group(function () {
     // B L O G
-    Route::post('/post', [PostController::class, 'create_post']);
-    Route::delete('/post/{id}', [PostController::class, 'delete_post']);
+    Route::post('/post', [PostController::class, 'create_post'])->middleware(['auth:api', 'role:SuperAdmin|Admin|SubAdmin|post a post']);
+    Route::delete('/post/{id}', [PostController::class, 'delete_post'])->middleware(['auth:api', 'role:SubAdmin']);
     Route::get('/post/{p?}', [PostController::class, 'get_post']);
     Route::put('/post/{id}', [PostController::class, 'edit_post']);
     // P E R M I S S I O N 
     Route::post('/create_role',[PermissionController::class, 'create_role']);
-    Route::post('/create_permission',[PermissionController::class, 'create_permission']);
-    Route::post('/create_subadmin',[PermissionController::class, 'create_subadmin']);
-    Route::get('/get_all_roles', [PermissionController::class, 'get_all_roles'])->middleware(['auth:api', 'permissionCheck:post a post']);
+    Route::post('/assign_permission',[PermissionController::class, 'assign_permission_to_role'])->middleware(['auth:api', 'role:SuperAdmin']);
+    Route::post('/create_subadmin',[PermissionController::class, 'create_subadmin'])->middleware(['auth:api', 'role:SuperAdmin|Admin']);
+    Route::post('/create_permission',[PermissionController::class, 'create_permission'])->middleware(['auth:api', 'role:SuperAdmin']);
+    Route::get('/get_all_roles', [PermissionController::class, 'get_all_roles'])->middleware(['auth:api', 'role:SuperAdmin|Admin']);
+    Route::post('/revoke_permission',[PermissionController::class, 'revoke_permission_of_role'])->middleware(['auth:api', 'role:SuperAdmin']);
+    
     
 });
 Route::get('/debug-sentry', function () {
